% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleBN.R
\name{downscaleBN}
\alias{downscaleBN}
\title{Downscale with Bayesian Networks}
\usage{
downscaleBN(DBN, x, prediction.type = "probabilities", event = "1",
  threshold.vector = NULL, output.attr.evidence = FALSE, cl = NULL,
  stop.cluster = TRUE, parallelize = FALSE, n.cores = NULL,
  cluster.type = "FORK")
}
\arguments{
\item{DBN}{Downscaling Bayesian Network, as returned by \code{build.downscalingBN()}.}

\item{prediction.type}{Options are: \code{"probabilities"}, \code{"event"}, \code{"simulation"} or
\code{"probabilities.list"}}
}
\author{
M.N. Legasa
}
