% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.starter.R
\name{parallel.starter}
\alias{parallel.starter}
\title{Parallelization Cluster Starter}
\usage{
parallel.starter(type, n.cores, PSOCK.funcExports.list = list(),
  PSOCK.varExports.list = list())
}
\arguments{
\item{type}{Cluster type, either FORK or PSOCK.}
}
\author{
M.N. Legasa
}
