% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDBN.R
\name{plotDBN}
\alias{plotDBN}
\title{Plot Bayesian Network}
\usage{
plotDBN(DBN, nodes = -1, node.size = 1, edge.arrow.size = 0.25,
  break.axis = 1, separation.ratio = 0.1, dev = FALSE)
}
\arguments{
\item{DBN}{Downscaling Bayesian Network, as returned by build.downscalingBN(). This is a wrapper for plot.restrictedgraph.R}
}
\author{
M.N. Legasa
}
