% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.BN.R
\name{downscale.BN}
\alias{downscale.BN}
\title{Downscale with Bayesian Networks}
\usage{
downscale.BN(DBN, x, prediction.type = "probabilities", event = "1",
  threshold.vector = NULL, output.attr.evidence = FALSE, cl = NULL,
  stop.cluster = TRUE, parallelize = FALSE, n.cores = NULL,
  cluster.type = "FORK")
}
\arguments{
\item{DBN}{Downscaling Bayesian Network, as returned by \code{build.downscalingBN()}.}
}
\author{
M.N. Legasa
}
