% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleBN.R
\name{downscaleBN}
\alias{downscaleBN}
\title{Downscale with Bayesian Networks}
\usage{
downscaleBN(DBN, x, output = "probabilities", prediction.type = "exact",
  event = "1", threshold.vector = NULL, output.attr.evidence = FALSE,
  cl = NULL, stop.cluster = TRUE, parallelize = FALSE, n.cores = NULL,
  cluster.type = "FORK")
}
\arguments{
\item{DBN}{Downscaling Bayesian Network, as returned by \code{build.downscalingBN()}.}

\item{output}{Options are: \code{"probabilities"}, \code{"event"} and
\code{"probabilities.list"}. You should probably not use the last one. Note that if
\code{prediction.type = "simulation"}, output is forced to \code{"event"}.}

\item{prediction.type}{Options are \code{"exact"}, \code{"approximate"} and \code{"simulation"}.
Exact inference requires a compilable junction.}
}
\author{
M.N. Legasa
}
