% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDBN.R
\name{plotDBN}
\alias{plotDBN}
\title{Plot Bayesian Network}
\usage{
plotDBN(DBN, title = NULL, dev = FALSE, nodes = -1, no.labels = FALSE,
  vertex.label.dist = 0.05, no.colors = NULL, node.size = 3,
  edge.width = 0.6, edge.arrow.size = 0.2, break.axis = 1,
  separation.ratio = 0.1, Nlabels = 4)
}
\arguments{
\item{DBN}{Downscaling Bayesian Network, as returned by build.downscalingBN().}
}
\author{
M.N. Legasa
}
