% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLatLonDAG.R
\name{plotLatLonDAG}
\alias{plotLatLonDAG}
\title{Plot a Graph placing nodes in euclidean coordinates \code{positions}}
\usage{
plotLatLonDAG(bn, positions, distance = NULL, nodes = -1,
  no.colors = FALSE, no.labels = FALSE, vertex.label.dist = 1,
  node.size = 4, edge.width = 0.5, edge.arrow.size = 0.15,
  edges.color = c("gray", "black"), dev = FALSE, axes = TRUE,
  xlab = "x", ylab = "y")
}
\author{
M.N. Legasa
}
